Bluetooth
Implement a proper way to ask for bluetooth permission at startup and whenever it is needed.
Also implement a proper way to monitor the bluetooth status.
Add custom alert dialogs to show a rationale for the bluetooth permission and off state.
Add an activity and a service with which the app will connect and communicate with the external bluetooth device.
Sort all files into new packages.
Create a temporary screen in main activity with which the bluetooth status monitor and permission requests can be tested.