Bluetooth feature
Add a fragment in which the user can see all paired devices and choose which device to use.
Add a fragment in which the user can see all available devices and pair with them.
Raw display data from HC-06 bluetooth module in sensor records fragment.
Remove DeviceControlActivity.
Stop using ble communication as the HC-06 bluetooth module does not support it.
Add data source in which the address of the primary device is stored.
Change StartupChecksActivity to check and ask for location permissions too.
Add new strings, colors and drawables.